Problems in set_wep and get_wep fixed.
get_wep should return the key only if the key is present. The status returned
is not a flag with all possible values, but an enum, so don't do bit operations
there.

set_wep should disable the key, wep, auth mode when a key is disabled.
Right now 'no key' is same as 'key disabled'. This needs to be fixed later.