More WEP fixes.
Change key only if one is given. Don't assume security mode to be either
open or restricted; check what is given in both cases. Assume restricted
mode if neither is specified, which seems to work for most people, but some
seem to need open mode; give a warning in those cases.