First cut at signal quality. The quality computed is rather vague; need to see how others have implemented it and see if we can improve it.
The data structure in NDIS uses 64 bit fields. However, there seem to be
no clean way of computing 64 bit division (when the divisor is 64 bits).
So for now we convert them to 32 bits (just chop them). This is not good,
but will do (what is the probability that someone will use wireless card
for __u32 to overflow - typical argument that may later prove wrong?).