Remove background ap scanning and make it on demand.
Backgroudn scanning apparently interferes with IP traffic. Instead, make
it on demand. But there has to be some delay between when scanning starts
and when the scan results are available. Right now, this delay is set
to 3 seconds. It seems wireless-tools don't like 5 second delay. We
have to find out if 3 seconds is good enough.

Couple of other minor reorganizations.