Various timers (ndis timer, ndis_miniport timer and Ke timer) are implemented with correct data structures as defined in NDIS.
Since none of these structures have the same space required by linux timer
support, we take up one of the fields in kdpc structure and allocate space
for linux timer there. This space is released when the driver is unloaded.

In terms of functionality, these changes don't change anything. But if a
new driver uses some of the fields in kdpc, for example, then these can
be easily implemented in future.