ndis_irq is now as per NDIS documentation, so instead of using it as a pointer, the fields are filled in properly.
disable_interrupts should be called while holding the spinlock, so it should
be in irq_th. Some checks are made (whether dynamic irq enabling/disabling
is used when the irq is shared etc.) to make sure the both ndiswrapper and
ndis driver are correct.

send_packets is called holding a spin lock. This seems to be not necessary,
but someone got an oops in send_packets. The real problem is probably not
in send_packets, as this lock doesn't help.