The due_time in KeSetTimerEx is in 100-nanosecond granularity, so conversion should use HZ/10000000.
When repeat is given for absolute timer, don't give error; it is okay to
give repeat for absolute timer (this means execute the first timer as
absolute timer and later timers as relative).

Ndis{Acquire/Release}SpinLock functions should not save/restore irqs. Watch
out for this one - according to document this should not mess with irqs. If
NDIS driver needs to do that, it should instead call
NdisMSynchronizeWithInterrupt, so no need to save/restore irq.