Many changes to locking! This has been tested with BCM94306, TI and Cetrino 2100 with 2.6.7 kernel. This version may not compile/work in 2.4 kernels.
Spinlocks in non-bh functions should not use bh versions. Moreover, use
bh versions only when the same lock is used by a bh function.

Spinlocks are now more verbose and extra checks are made to make sure
the locks are initialized properly. These were removed in 0.8 hoping that
our spinlock implementation wrt NDIS is correct. But it looks like we
are not there yet!

Generic workqueus are replaced with ndis_wq workqueue to avoid having
to call schedule_work. This seems to avoid 'scheduling while atomic'
errors in some cases, although not sure if this is the correct way to
handle it.

NdisWaitEvent should return immediately if the event state is already set.

When reset is done, check if address and other filters need to be initialized
and do so if required.

Don't use same spinlock in xmit_bh and sendpacket_done. At least Centrino
2100 driver calls NdisMSendComplete from within send_packet. Consequently,
NdisMSendComplete needs the same spinlock held in xmit_bh, which is not
correct. So we now have two separate spinlocks.