Linux spin lock functions are replaced with wrapper functions. These wrapper functions check if the calls are in atomic context or not. If they are, then non_bh version of spin_lock is used, otherwise, _bh version is used. With these wrapper functions, it is easy to change this policy easily at one place instead of all over the code.
NdisFreeBufferPool shouldn't call flush_scheduled_work, as it may be called
in atomic context. Pontus thinks it is necessary to do it, but without it,
everything seems to work fine (with three different drivers). If it is indeed
needed by a driver, we need to think of alternate way of implementing it.