Set authentication and encryption modes in separate functions so the current mode can be stored in handle. This way we know for sure what is the current mode by looking in handle. This is used in, e.g., wrapper_worker_proc.
When resetting, if the driver indicates that address and filter need to be
reset, do so only doreset. This also simplifies implementation of both
doreset and NdisMResetComplete.

Similarly, dosetinfo and doquery have been simplified and unnecessary goto
eliminated.

If connection status changes, find out association information only when
WPA is active. Some drivers apparently crash when querying association
information (only if WPA is not active?).

Don't call doreset in ndis_init_one if it is Centrino card (apparently
these cards don't associate if the card is reset). Not sure if this problem
is only with Centrino 2200 or 2100 too.