Split functions in iw_ndis.c into two groups: ones called by wireless tools (these are prefixed with iw_) and ones used internally by ndiswrapper (these are without prefix).
Support for ciphers other than WPA-PSK+TKIP is added, including WEP. So it
is now possible to use wpa_supplicant for setting WEP keys.

When setting WEP through wpa_supplicant, convert the keys in ascii to hex.
Use iw_set_encr to set the keys instead of doing the same thing all over
again in wpa_set_key.

When checking for highest encryption supported, check if querrying later
returns key absent, which actually means that mode is supported but keys
are not available yet. This way we don't force the driver to use TKIP, but
the highest possible encryption.