Call usb_init when module is loaded; this initializes variables in usb layer.
Releasing cancel_lock in usb_cancel_transfer doesn't seem to be correct; if
the only path to get there is from IoCancelIrp, then this can be done
right before calling cancel_routine. Otherwise, it is not correct to release
lock here, as there seem to be no matching lock elsewhere in usb.c/ntoskernel.c

With these, wusb54g driver seems to work with SMP and PREEMPT separately
(tested on UP machine, so it is still possible that there are problems on
SMP machine). However, using both SMP and PREEMPT at the same time crashes
kernel. Again, this test is done on UP machine.