This large patch adds probably the last feature for 1.0 release.
Some drivers have many devices (e.g., Centrino). Loading all devices, along
with the settings is not necessary. Moreover, it is not necessary to load
all drivers when one doesn't need to use them. This patch implements dynamic
driver/device loading. Initially all devices are registered and when a device
needs to be configured, settings and driver specific to that device are loaded
automatically. Drivers are not unloaded automatically, though.

One must re-install Windows drivers now, as the configuration file names
now include the bustype so we don't have to parse the config file when loading
devices.