Use spinlocks whenever invoking a miniport function with appropriate IRQL. This may fix SMP issues in some cases.
Remove pe_linker.h, as it is not necessary anymore.

When timeout is positive, subtract ticks_1601 from it. Then check if the
value is negative. It seems some drivers use very low value which indicates
either drivers are not using what NDIS says or NDIS spec is not correct. Anyway,
so far the USB drivers that use this seem to be working with this.

X86_64 may have either PAGE_KERNEL_EXECUTABLE or PAGE_KERNEL_EXEC; use whichever
one is available; convert types to unsigned long instead of unsigned int when
a variable can be either; no need to use 64-bit/32-bit fields of physical
addresses separately; these three patches are from Frank Zago.