It seems when scatter/gather DMA is enabled and more than one buffer is sent, RTL8180L driver slows down. If only one buffer per packet is sent (effectively disabling the benefits of SG), it works very well. So send only one buffer per packet. Other drivers may work well when sending more than one packet.
Protecting each miniport function with a spinlock at DISPATCH_LEVEL
was introduced recently in hopes that might fix some SMP issues. Ideally,
ndiswrapper shouldn't need to do this - the Windows driver should obtain
spinlock when necessary. Someone reported that SMP works with earlier
version but not now, so reverting to not using spinlock, but just IRQL
when calling miniport functions.