Implemented NdisBuffer allocate/free with pool. This avoids repetitive allocating/freeing buffers. Ditto for packets.
There is one caveat. A buffer may be allocated for one region with a length
which may be different for a different region with different length. As per
NDIS, the array at the end of the buffer decriptor should have enough space
for length of region. However, in ndiswrapper these buffers don't describe
the region the way in Windows. Since the buffers are opaque to the driver,
a sane driver shouldn't care about this. If a certain driver does care, it
may break. One approach to implement this so that each buffer has enough
space for every region, we could always allocate space for biggest possible
regionn. Another one is to allocate buffer for regions of a certain length
and if a buffer is requested for a region that is more than this length, treat
it differently (e.g., allocating from heap). However, at least for now, I
don't see anything wrong with just not caring about the length.

Implemented new status types in NdisMIndicateStatus (e.g., radio status).
However, no driver seems to use it (as yet), so this may not be useful now.

When allocating packets in EthRxIndicateHandler, don't use NdisAllocatePacket
and NdisFreePacket, as we don't have a handle to pool. Instead, allocate/free
with kmalloc and kfree. These cases are rare, if at all they occur, so no
harm with that.

Moved debug module parameter and initialization into wrapper.c.

Call usb_cancel_transfer holding irp_cancel_lock spinlock, as per Windows.