Thanks to Bill Paul (of Project Evil) for pointing out grave bug: In previous implementation, we used maps to spinlocks and the driver should have called NdisAllocateSpinLock before calling NdisAcquireSpinLock, but TI ACX 100 driver doesn't, so we were calling NdisAllocateSpinLock if there was no mapping for it. In the current implementation, NdisAllocateSpinLock, it resets the lock to UNLOCKED state, which means NdisAcquireSpinLock wouldn't work as spinlock. Bill pointed out the mistake, which saved lots of trouble and time.
NdisDprAcqireSpinLock should call kspin_lock at PASSIVE_LEVEL, so IRQL is
not changed. Earlier it was being called with DISPATCH_LEVEL so that if a
driver is called this function at PASSIVE_LEVEL, we raise the IRQL as
necessary, but that would be against the semantics of NdisDprAcquireSpinLock.
In deference to Bill Paul (especially after the above notice), we now do
the way he prefers it (to follow semantics of NDIS and hope that the driver
does the right thing).