This patch adds many functions used by Conexant modem. However, it looks like this effort is not going to be useful, thanks to Microsoft not getting the kernel/user space issues right - try/exception blocks, creating files and conversion to calendar time in kernel space? While I was franatically implementing the functions, I came across ProbeRead, which raised suspicions that the driver might be using 'fs' register for exceptions and sure enough, the driver does. There is no way this can work in Linux (or should in any decent kernel). It may be possible to just 'nop' those instructions in the driver and get it to work, but chances are low this may work. So for now this effort of getting modem driver to work is put on hold, if not abandoned.
While implmenting semaphores, it occurred to me that dh->signal_state is
actually counter (or at least it can be used as such), so instead of using
it as boolean variable, we now use it as a counter. Object types mutex,
events initialize it to 1 when in signaled state and decrement it (to 0) when
not in signaled state.

None of these functions (including functions that have been modified) have
been tested.