Started DEVEL branch for development so it is easier to manage development so changes won't affect main branch.
This patch fixes two serious races:

* In KeWaitForMultipleObjects, we should check if a thread can grab the
event with check_reset_signaled_state before putting a thread on the wait
list and sleep. Once an is set, wakeup_threads is called to wake up any
threads waiting on that event. This will go through the list of waiting
threads, remove one from the head and atomically check and grab the event
on behalf of the thread removed. If this call succeeds, the thread is woken
up, which no longer (again) check and remove from wait list (which was
happening until now). We use wb->object to indicate that wakeup_thread
grabbed the event on behalf of the thread and the thread should not check
again if it can grab the event. This fixes the race where, (a) a thread
on the wait list is not removed twice (once by wakeup_threads and once by
the thread itself) and (b) an event is grabbed and possibly is reset to
non-signaled state in one atomic operation.

* In USB layer, once a packet is submitted for transmission, Linux USB
layer owns the packet until the completion function (usb_transfer_complete)
is called. However, in the meantime, driver may want to cancel the packet
with IoCancelIrp. If IoCancelIrp is called at < DISPATCH_LEVEL, we call
usb_kill_urb right away and the packet will most likely be killed (there
is a potential for race here still - read below). However, if IoCancelIrp
is called at DISPATCH_LEVEL, we can't call usb_kill_urb at that point (since
usb_kill_urb can be called only in non-atmoic context). So we put the urb
on urb_cancel_list and schedule a worker thread to process it later. Before
this worker thread can work on the packet, however, it is very likely that
the Linux USB layer may finish transmitting the packet and call the completion
function, which puts the _same_ packet on the urb_tx_complete_list and
calls its worker thread. Then the packet is seen on both lists and both
worker threds try to free the packet with wrap_free_urb! To avoid this race,
we first initialize the urb's list member so it is both successor and
predecessor of itself. If packet goes through successful transmission path
(without a call to IoCancelIrp), everything goes fine. If IoCancelIrp is
called, this packet is put on the urb_cancel_list and worker is scheduled.
If this worker removes the packet before the completion function is called,
again, everything goes fine. However, if completion function is called
before packet is killed, the completion function first removes the packet
from list (with RemoveEntryList): If packet was never on the cancel list, this
call does nothing and if packet is on the cancel list already, it is removed
from that list. Then it puts the packet on completion list and its worker
is scheduled - the worker for cancel list never sees this packet. All this
problem is because we are killing packets asynchronously, so driver and
Linux USB layer may see the packet in different states. It is possible to
kill the packet right away in atomic context with USB_ASYNC_UNLINK flag, but
generally this is to be avoided (if my understanding of USB layer is correct).
It may be that we should just do that - we need testing to see what problems
this may raise.