As explained in previous commit, putting urb's on cancel list when IoCancelIrp is called is problematic - there could be a race condition where packet is queued for canceling but before it is killed, Linux USB layer transmits it. To avoid this, we now check if we can kill the urb synchronously (if IoCancelIrp was called in non-atomic context) or asynchronously (otherwise). Moreover, when a urb is canceled, we don't do IoCompleteRequest for this - the driver doesn't (or so I understand) expect to be called back. With these two fixes, unplugging the device doesn't crash anymore (with either rt2500usb or prisma02). However, prism1 driver which used to work earlier is no longer working reliably, but this could be due to other changes that went in other layers.
Fixed check_reset_signaled_state so if a thread owns the mutex, we should
decrement dh->signal_state even it is <= 0. Also added some debug checks
to make sure that logic is correct.

Some code in setup_device is commented out for now; with this ZyDas chipset
initializes the device, scan works, but won't associate (yet).