Process various vendor/class requests and set the request type for urb. This fixes initialization issue with SIS163 driver: Now the device sends and receives all urb's without errors and shows its MAC address correctly. However, scanning doesn't work yet.
All urb's submitted asynchrounously are now put on a work queue and processed
later. This is to avoid the race condition I mentioned earlier: After urb
is submitted and before we return STATUS_PENDING to the driver in
pdoDispatchDeviceControl, the urb is done and its completion function calls
IoCompleteRequest, and by the time STATUS_PENDING is processed, urb and
irp are already freed. To avoid this race condition, the work queue is
scheduled just before returning STATUS_PENDING in pdoDispatchDeviceControl.

There are also lot of changes here that were being committed to HEAD for
sometime.