Get rid of 'task_t' and use 'struct task_struct' in its place, as even some old kernels have backported its definition and it is not easy to check if it has been defined or not.
RT2500 USB driver needs to mirror urb buffer - looks like it just uses
urb on stack at high addresses.

Support for MDLs in URBs added, but this has not been tested and likely
MDLs are not used by any drivers that we may encounter.

Maintain status of urb with a flag so we know when/if to remove it from
tx_submit list when canceled etc.

Don't touch irp/urb in case urb has been queued for submission, as it may
be done before we check the status in which case irp/urb are invalid. Instead,
use return value from various functions that submit urb.