In pdoDispatchInternalDeviceControl, don't access irp if STATUS_PENDING is returned. This avoids race condition mentioned in previous commit. The status of IRP is set in usb_submit_urb and usb_reset_port.
Yet more tweaking of canceling of timers. The description for canceling of
timers is not crystal clear, so not sure if current implementation (or
previous one) is correct, but with this implementation new Marvell driver
(in Netgear WG511v2) sets essid without timing out.

Implemented Io(Acquire|Release)CancelSpinLock for processing IRPs.

When an IRP is canceled, complete it with appropriate value in IRP and URB
instead of just ignoring it - earlier I thought if driver cancels IRP,
it doesn't expect to get it back.

IRP processing within USB layer needs redesign. I just discovered a race
condition: usb_submit_nt_urb will submit urb and in the case of
bulk/interrupt or class/vendor urb's, may get the status from their
processing functions that the urb is pending. Before usb_submit_nt_urb
returns, however, urb is done and the completion function for urb calls
IoCompleteRequest which frees urb and irp. Then usb_submit_nt_urb returns
pending status back to pdoDispatchInternalDeviceControl which again
submits irp for completion. We need to eliminate this double processing
of irps. This happens with RT2500 USB driver, although not always.

'Maximum send packets' is not used for deserialized drivers.

Provide usb_buffer_alloc and usb_buffer_free for 2.4 kernels - these are

Get rid of 'task_t' and use 'struct task_struct' in its place, as even
some old kernels have backported its definition and it is not easy to
check if it has been defined or not.

RT2500 USB driver needs to mirror urb buffer - looks like it just uses
urb on stack at high addresses.

Support for MDLs in URBs added, but this has not been tested and likely
MDLs are not used by any drivers that we may encounter.

Maintain status of urb with a flag so we know when/if to remove it from
tx_submit list when canceled etc.

Don't touch irp/urb in case urb has been queued for submission, as it may
be done before we check the status in which case irp/urb are invalid. Instead,
use return value from various functions that submit urb.

Use usb_buffer_alloc to provide DMA buffers to transfer data. Use Linux's
endpoint as reference in pipe_handle, instead of creating pipe information.
The only field that we can't store with this is max_tx_size, but we are
not using it (at least right now). RT2500 USB driver works (again) with
these changes.

mm-series kernels use PCI_D0 etc for suspend/resume/enable to PCI devices,
so use (and define for older kernels) those symbols. Change suspend/resume
routines as necessary. This hasn't been tested on older kernels, so not
sure if this breaks on old kernels.

Apparently 'halt' is a macro in mm-kernel, so use miniport_halt instead.

Asynchrounous unlinking of URB's is also gone from mm-kernels, so get
rid of it.

All functions that are used only in USB layer are declared static.

In USB_CreateConfiguratiomRequestEx, initialize 'intf' fields of select_conf
with corresponding field in intf_desc in the given intf_list. Allocate space
for these fields and initialize them zero'ing out so wrap_select_conf knows
which fields are initialized (only intf->bLength is checked for validity).
With this prism54 driver works (again).

Further (end of) cleanup of USB layer:

USBD_(Create|Parse)Configuration functions are (re)implemented as per
their description. Specifically CreateConfiguration function should
allocate space for nt_urb with enough space for interface information
and initialize pipe information there, so that the driver can fill in
information in that structure before selecting desired configuration.

Store information about pipes within wrapper_dev instead of encoding
it in a pointer.

None of this implementation is tested.

Process various vendor/class requests and set the request type for urb.
This fixes initialization issue with SIS163 driver: Now the device sends
and receives all urb's without errors and shows its MAC address correctly.
However, scanning doesn't work yet.

All urb's submitted asynchrounously are now put on a work queue and processed
later. This is to avoid the race condition I mentioned earlier: After urb
is submitted and before we return STATUS_PENDING to the driver in
pdoDispatchDeviceControl, the urb is done and its completion function calls
IoCompleteRequest, and by the time STATUS_PENDING is processed, urb and
irp are already freed. To avoid this race condition, the work queue is
scheduled just before returning STATUS_PENDING in pdoDispatchDeviceControl.

There are also lot of changes here that were being committed to HEAD for
sometime.

Wait for 2 seconds instead of 1 second in MiniportQueryInformation and
MiniportSetInformation; otherwise, prism1 USB driver crashes (possibly
because it uses the buffers passed in those functions after these
functions timeout) when wep key is set.

Prism1 driver for MA111 from Netgear seems to need longer timeouts for
Miniport(Query/Set)Information functions (with 4 seconds it seems to
work fine). However, just because of one driver, requiring all the
other drivers to wait for same seems insane. So added a field
ndis_comm_wait_time to wd structure which is initialized depending on the
device (for prism1 devices DWL-122 and MA111 it is 4 seconds for others
it is 2 seconds).

'Maximum send packets' is not used for deserialized drivers.

Provide usb_buffer_alloc and usb_buffer_free for 2.4 kernels - these are
not tested and DMA is not used.

Use usb_buffer_alloc to provide DMA buffers to transfer data. Use Linux's
endpoint as reference in pipe_handle, instead of creating pipe information.
The only field that we can't store with this is max_tx_size, but we are
not using it (at least right now). RT2500 USB driver works (again) with
these changes.

Submit bulk/interrupt/vendor request URBs asynchrounously so memory for
URBs can be allocated with GFP_KERNEL (some drivers submit USB requests
in atomic context).

Maintain various states of URB (submitted, canceled, pending in queue etc)
in a variable in irp so we can process URB appropriately. Looks like there
is something wrong in maintaining this flag - with ZyDas driver URB state
is wrong sometimes.

When URB is done, copy data back to NT URB's buffer if necessary (if DMA
buffer was allocated earlier for this).

At this point, ZyDas driver initializes the card, scans stations, but
miniport_set_info with encryption, mode etc. hang. SIS163 driver
subimts/receives many URBs successfully, but later fail with -EILSEQ -
I am not sure if this is some problem with the hardware, as the kernel
complains about reset and initializes it as full speed device, instead of
high speed device (ehci-hcd module can't initialize properly, so uhci-hcd
initializes it as full speed device only). It is possible these initialzation
issues are causing URBs to fail.