Simplified implementation of timers: Use return value of del_timer to know if timer is active, instead of maintining active flag, use mod_timer to either add or modify timer, add timer without using timer_lock etc.
When canceling timers from misc_funcs_(device_)exit, don't access ktimer
field of wrap_timer - it is likely the driver has already freed the timer
without canceling it.

When submitting urb's to Linux kernel, for asynchronous submission (bulk/int
or vendor/class), delay submission and let pdoDispatchControl do it. This
is to prevent the race condition (where urb's completion function may return
the irp back to driver and then pdoDispatchControl returns STATUS_PENDING)
as much as possible. With preemption, it is still possible for this race,
and I have no idea how to avoid it completely. With this, we don't need
worker threads to process urb's: they are processed right away.