Merging changes that went in HEAD.
When allocating NDIS packet pool, consider ProtocolReservedLength, which
indicates extra bytes that need to be allocated for each packet allocated
from that pool. Although DDK states it must be 4 * sizeof(void *), the
header files indicate that this is the minimum and a driver may request
a bigger size. So we now store this value in the pool handle and whenever
a packet is allocated, we add this to the length (well, actually a few more
bytes, to be specific). The packets are no longer allocated with kmem cache
as each packet pool may have different length due to different values for
ProtocolReservedLength. It is possible to allocate a unique kmem cache
for each packet pool, but that may be too much of overhead. Instead, packets
are allocated with kmalloc and freed back to the packet pool instead of with
kfree, upto a point (MAX_ALLLOCATED_NDIS_PACKETS); after that packets are
freed back to kernel. This will keep a balance of packets that we hold onto
hoping to give them back to driver without allocating/freeing too much and
not keeping too many packets in the free chain.

With the above change, a separate packet pool is allocated for use by
ndiswrapper to send packets to the driver.

Instead of allocating URB's whenever driver requests an IRP submission,
certain number of URB's are allocated when a device is initialized in
usb_init_device. With this, it is easy to manage different states of
URB as well as to cancel / free URBs when required (we may need to cancel
URBs when suspending device and resubmit them when resuming). Right now
we allocate only 10 URBs hoping that is the maximum number of URBs that
the driver needs, but some drivers may need more.

Since both these above changes are critical, this may require more time
for testing, which means some more time before 1.4-rc1 is released.