Separate NDIS and Ke timers; set only Ke timers to signal state, as DDK doesn't say anything about signaling events for NDIS timers.
When initializing NDIS timers, don't use KeInitializeTimer and then
wrap_init_timer - that creates two wrap_timers.

When sending packets, allocate NDIS buffers from a separate pool and create
and destroy this pool.

In wrap_urb structure, use a field to note if a buffer has been allocated
for URB or not, so it can be freed when URB is put back on the free list.

Implemented functions to cancel URB during suspend and submit during resume.
These are not used yet and not even sure if these are required - drivers
should cancel IRPs, but these routines may help to fix if a driver doesn't
do that.

Now Airgo Pre-N driver doesn't bark when setting WEP, which it used to
do earlier. But when sending packets, it still does, apparently right
after NdisQueryBufferSafe returns.