Merging in changes that went in DEVEL branch. The following are the changes:
Simplified implementation of timers: Use return value of del_timer to
know if timer is active, instead of maintining active flag, use
mod_timer to either add or modify timer, add timer without using
timer_lock etc.

When canceling timers from misc_funcs_(device_)exit, don't access
ktimer field of wrap_timer - it is likely the driver has already freed
the timer without canceling it.

When submitting urb's to Linux kernel, for asynchronous submission
(bulk/int or vendor/class), delay submission and let
pdoDispatchControl do it. This is to prevent the race condition (where
urb's completion function may return the irp back to driver and then
pdoDispatchControl returns STATUS_PENDING) as much as possible. With
preemption, it is still possible for this race, and I have no idea how
to avoid it completely. With this, we don't need worker threads to
process urb's: they are processed right away.

Get rid of the variables to put urb's on a submit/complete lists since
we don't use workers anymore.

Remove 'static' classifier for exported Windows functions from usb.c.

Remove debug code to signal events and threads from proc.c.

Use nt_list_for_each_entry instead of nt_list_for_each for simplicity.

Timers and DPCs are simplified - until now the major pain of dealing
with kernel timers and NDIS timers, kernel DPCs and NDIS DPCs was due
to confusing explanation for canceling of timers (e.g., KeCancelTimer)
which seems to indicate that DPCs must be dequeued when canceling
timer. However, elsewhere on MSDN, it seems to say KeCancelTimer
should return TRUE if timer is pending and FALSE otherwise. When FALSE
is returned, it means either timer is not pending or DPC is executing
(or executed). This has been confirmed with developers at #reactos. So
all that work of keeping timers and DPCs of kernel and NDIS has been
removed. So far the drivers that have been tested seem to work fine
with this change. If any other drivers fail, this fix may have to be
investigated.

It seems Pre-N (Airgo) driver is quite buggy wrt timers: It sets timer
with NdisSetTimer and cancels with NdisMCancelTimer (instead of
NdisCancelTimer).  More annoyingly, it initializes a timer with
NdisSetTimer before initializing it with NdisInitializeTimer. So we
now check if a timer has been initialized or not in wrap_set_timer and
initialize it if not. There seem to be some issues with
packets/buffers with this driver, too - this has to be investigated.

Semaphore objects are initialized now as NotificationEvent instead of
SynchronizationEvent and when a thread wants to grab it, we check
signal_state and decrement it if it is signaled.

Add missing functions for usb support in 2.4 kernels
(usb_buffer_alloc, usb_buffer_free etc.). All this has been only
compile-tested only - not sure if any of this really works.

Use KTHREAD_RUN macro to support software suspend2.

For some reason the fix for prism1 usb driver doesn't work with
Broadcom driver: If expires == 0, executing the dpc right away instead
of adding timer seems to cause problems with Broadcom driver (this
doesn't make any sense). So kdpc's are divided into two types: kernel
and NDIS. Whereas kernel kdpc's are put in dpc queue, NDIS kdpc's are
executed when the timer expires. The type of kdpc should be
initialized everytime a timer is set, not just at initialization time,
as a driver can associate a kdpc that was initialized for kernel timer
can later be associated with NDIS timer.

Use worker to process completed urb's, as IoCompleteRequest can
sleep. Don't refer to urb->dev once the urb is completed, as it may
have been cleared by usbcore.

In wrap_free_urb, initialize urb with usb_init_urb so it can be given
to another irp later without any hangovers.


Separate NDIS and Ke timers; set only Ke timers to signal state, as
DDK doesn't say anything about signaling events for NDIS timers.

When initializing NDIS timers, don't use KeInitializeTimer and then
wrap_init_timer - that creates two wrap_timers.

When sending packets, allocate NDIS buffers from a separate pool and
create and destroy this pool.

In wrap_urb structure, use a field to note if a buffer has been
allocated for URB or not, so it can be freed when URB is put back on
the free list.

Implemented functions to cancel URB during suspend and submit during
resume.  These are not used yet and not even sure if these are
required - drivers should cancel IRPs, but these routines may help to
fix if a driver doesn't do that.

Now Airgo Pre-N driver doesn't bark when setting WEP, which it used to
do earlier. But when sending packets, it still does, apparently right
after NdisQueryBufferSafe returns.

Use tasklet instead of worker to process completed URBs. Apparently
with ZyDas driver the worker never gets executed after a while; right
now I don't know why this happens, but for now, I will take the easy
route and use tasklet. With this, ZyDas driver seems to work nicely,
at least with WEP - WPA hasn't been tested yet.

Disable wireless stats collection for ZyDas: If it is enabled,
Miniport Set/Query functions take forever to finish; this causes
wpa_supplicant to fail as it doesn't get association information in
time.

When device is opened, set packet filter appropriately. When setting
multicast list, update packet filter appropriately.

With these changes, ZyDas works with WPA, albeit with TKIP only;
although the device claims AES capability, I had trouble associating
with with this - it may be problem specific to my setup.

Allocate NDIS buffers and NDIS packets ignoring maximum number of
descriptors requested, except to give warning; allocated descriptors
are chained in free list so a new descriptor is fetched from that
list, instead of allocating afresh. When a descriptor is freed, put it
in free list if number of allocated descriptors is below a limit,
otherwise free it back to kernel.

When converting time units to jiffies for expiration timers, add 1, so
that jiffies is rounded up, instead of down.

KePerformanceCounter should set counter to HZ, not to jiffies.

After even more debugging, found out that ZyDas driver crashes when
ASSOCIATION_INFORMATION oid is called. This is probably because we
don't fill the information as expected by the driver. DDK seems to say
that at least some of the fields in there should be initialized with
earlier request or something similar to that. I don't quite understand
the description about it, so for now, we just don't request it for
ZyDas driver.  I thought with this wpa_supplicant doesn't work, but
thankfully it does.

With these changes, sis162 also works (someone confirmed earlier that
sis163 works too). So it is time for 1.4-rc1.