Moved timer functions from misc_funcs.c into ntoskernel.c as all timers are now implemented using kernel timers.
NdisFreePacket was not releasing spinlock when if condition there
succeded. This fixed errors with rmmod'ing with sis162 driver. This may
also fix the same problem with ipn2220 64-bit driver.

Call KeFlushQueuedDpcs when a device is being removed so any pending
DPCs get called.

Adding 1 to jiffies when converting timers is wrong, when used for computing
repeat duration etc when the result otherwise could be 0. Morever,
SYSTEM_TIME_TO_HZ should consider the case when sys_time is 0.