Calling miniport_pnp_event_notify after miniport_halt apprently causes crash. So instead of sending stop and remove irps when a device is removed, send only remove device and call miniport_halt for both stop and remove irp handlers in wrapndis.
Someone claimed that older version of ndiswrapper works with 4k stack but
not the newer ones. This _may_ be because we use tasklets instead of work
queues for miniport irq handler. The information on msdn is not at all clear
about conditions for calling isr, so for now reverting to work queues.