Importing all the changes that went into HEAD branch.
This branch differs from HEAD branch in implementation of IRQL related
functions. It implements DEVICE_LEVEL to handle the case when irqs are/can be
disabled. We also use IRQL to "remember" the state so that it is safe to
enter DEVICE_LEVEL from DEVICE_LEVEL etc.

With this patch, prism54 usb driver works.