During 1.11, Rtl functions have been modified to get Am1772 driver working. We now don't assume given Rtl functions are not counted ('length' feild may not be correct) and count the given string when converting. Use the counted length to check if conversion can succeed or not.
TI 1450 USB driver (for AVM Fritz) requires the USB device to be reset fully
when initializing. Otherwise, if module is reloaded, device fails to initialize.
This reset probably can be done for all devices, but for now, do it only
for this device.

When infrastructure mode is changed, NDIS drivers reset the keys. However,
Linux wireless tools assume otherwise. So restore the keys when changing
the mode, so wpa_supplicant etc work. With this, WEP with 'wext' driver
interface through wpa_supplicant works. Since 'wext' is fully supported now,
'ndiswrapper' interface is no longer needed and marked obsolete - a warning
message is given if 'ndiswrapper' interface is used.

To deal with kernel crashes, create a separate work queue which would otherwise
be default workqueue. This can be enabled/disabled with USE_OWN_NTOS_WORKQUEUE
macro in ntoskernel.h. By default this is disabled (default workqueue is used).

Now that NT thread / Linux tasks are handled appropriately (some places
we use NT thread and at others we use Linux task), convert warning messages
to debug messages.

Sis163 driver needs 26 more bytes for association information, so allocate
memory accordingly.

Disable power management when initializing the card (it can be enabled with
iwconfig's 'power' command or through private ioctl).

io_status_block is defined as per DDK. With this crashes with 64-bit USB
driver fixed (tested with AVM Fritz). But the driver fails to initialize the
card and there is no clue why this happens. Other 64-bit USB drivers may
work.

When usb device is being reset fully in wrap_reset_port, lock the device
for reset. We probably need to reconfigure the device, but documentation is
not clear about this.