Atheros USB driver seems to corrupt the memory past the end of ndis_miniport_block - freeing net_dev results in slab error indicating memory corruption. Apparently keeping a barrier of 20 bytes seems to work. I am not sure if ndis_minport_block used by Atheros and ZyDas is different or somehow they are compiled with NDIS_WRAPPER enabled, which adds quite a few fields to ndis_miniport_block.
Just noticed that ZyDas driver works now with wpa-psk with tkip. Till now,
I could get it to work with wpa2-psk wiht aes only.

Buffalo WLI-U2-KG54-AI (rt2500 usb) has a strange behavior: After using
it with wpa-psk (with wpa_supplicant), which resets encryption and
authentication modes at the end (to 1 and 0 respectively), if iwconfig
is used to set wep settings by first setting key and then essid, the
driver sets authentication and encryption modes to wpa-psk! If key is
turned off with iwconfig at this point, remove_key is called with bssid
as NULL, but since encryption mode is for wpa-psk, it will access bssid,
resulting in NULL pointer dereference. So bssid should be checked for
NULL to handle this case. Setting the key again with iwconfig sets these
correctly (again).

To disassociate, use OID_802_11_DISASSOCIATE oid; but this also disables
radio, so set random ssid to enable the radio again.