Show memory allocation information when a device is removed.
Atheros USB driver (or ndiswrapper with this device) leaks 124 bytes
allocated in atomic context each time it is used. Hopefully this can
be tracked down (and it is bug in ndiswrapper and not in Windows driver).

For now enabling ALLOC_INFO so whoever uses svn will find this information
useful and holler if memory is being leaked with a driver.

As stated before, indeed, removing free_netdev macro seems to work with
ZyDas and Atheros USB drivers without barrier in ndis_miniport_block.