This patch fixes issues with .inf parsing of Agere driver. Patch from Pavel Roskin. Below are his comments about the patch:
Following patch fixes support for Agere drivers for Orinoco cards:
http://www.agere.com/mobility/docs/windows_drivers_sr02-2.3.zip

All of those changes should be considered as fixes for bug that can
affect other drivers.

The worst bug by far is incorrect interpretation of the parseVendor()
result code in parseMfr(), which causes only the first vendor entry to
be processed!  I'm assuming that the code is supposed to exit if
parseVendor() failed and keep going if it succeeded.  Maybe it's better
not to exit in either case, but to keep going.

%fuzzlist doesn't keep the bus number.  Instead, the global $bus
variable is used, which caused broken links.  The patch introduces a
separate hash %buslist that keeps the bus data for every device
referenced in %fuzzlist.  Maybe %fuzzlist and %buslist should be
integrated into one array, but I'm trying to keep my patch simple.

The "Manufacturer" section contains quoted names on the right hand side.
The quotes are removed after splitting the string.  Fortunately, there
are no commas inside the quotes, but ideally, quoted commas should be
ignored.

In the vendor sections, the ID needs to be subjected to string
substitution.